Extract `CrossfilterOptions` component
